Split up resnet_test.rb into resnet_hers_test.rb and resnet_other_test.rb
